pyproject.toml
[tool.poetry]
name = "rag-nawfal"
version = "0.1.0"
description = ""
authors = ["Nawfal <nawfal.adil@telecom-paris.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
sentence-transformers = "^4.1.0"
datasets = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
